+++ packages/edge-login-ui-rn/src/components/abSpecific/FourDigitComponent.js
@@ -5,14 +5,14 @@ import { Text, View } from 'react-native'
 import { sprintf } from 'sprintf-js'
 
 import s from '../../common/locales/strings.js'
+import * as Colors from '../../constants/Colors.js'
 import * as Constants from '../../constants/index.js'
 import { type Dispatch, type RootState } from '../../types/ReduxTypes.js'
+import { scale } from '../../util/scaling.js'
 import { Spinner } from '../common/Spinner.js'
 import { connect } from '../services/ReduxStore.js'
 
-type OwnProps = {
-  style: Object
-}
+type OwnProps = {}
 type StateProps = {
   pin: string,
   autoLogIn: boolean,
@@ -49,14 +49,13 @@ class FourDigitComponent extends Component<Props, State> {
   }
 
   render() {
-    const Style = this.props.style
     return (
-      <View style={Style.container}>
-        <View style={Style.interactiveContainer}>
-          {this.renderDotContainer(Style)}
+      <View style={styles.container}>
+        <View style={styles.interactiveContainer}>
+          {this.renderDotContainer()}
         </View>
-        <View style={Style.errorContainer}>
-          <Text style={Style.errorText} numberOfLines={2}>
+        <View style={styles.errorContainer}>
+          <Text style={styles.errorText} numberOfLines={2}>
             {this.props.error}
           </Text>
         </View>
@@ -64,12 +63,7 @@ class FourDigitComponent extends Component<Props, State> {
     )
   }
 
-  renderCircleTest(style: Object) {
-    return style
-    // return {...style, borderColor: this.state.circleColor}
-  }
-
-  renderDotContainer(style: Object) {
+  renderDotContainer() {
     const pinLength = this.props.pin ? this.props.pin.length : 0
     if (this.props.wait > 0) {
       return <Spinner />
@@ -78,33 +72,61 @@ class FourDigitComponent extends Component<Props, State> {
       return <Spinner />
     }
     return (
-      <View style={style.dotContainer}>
-        <View
-          style={[
-            this.renderCircleTest(style.circle),
-            pinLength > 0 && style.circleSected
-          ]}
-        />
-        <View
-          style={[
-            this.renderCircleTest(style.circle),
-            pinLength > 1 && style.circleSected
-          ]}
-        />
-        <View
-          style={[
-            this.renderCircleTest(style.circle),
-            pinLength > 2 && style.circleSected
-          ]}
-        />
-        <View
-          style={[
-            this.renderCircleTest(style.circle),
-            pinLength > 3 && style.circleSected
-          ]}
-        />
+      <View style={styles.dotContainer}>
+        <View style={[styles.circle, pinLength > 0 && styles.circleSected]} />
+        <View style={[styles.circle, pinLength > 1 && styles.circleSected]} />
+        <View style={[styles.circle, pinLength > 2 && styles.circleSected]} />
+        <View style={[styles.circle, pinLength > 3 && styles.circleSected]} />
       </View>
     )
+  }
+}
+
+const styles = {
+  // used for logging *back in* with PIN
+  container: {
+    paddingTop: 12,
+    width: '100%',
+    height: scale(86)
+  },
+  errorContainer: {
+    height: scale(40),
+    width: '100%',
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-around'
+  },
+  errorText: {
+    color: Colors.ACCENT_RED,
+    backgroundColor: Colors.TRANSPARENT,
+    textAlign: 'center',
+    fontSize: scale(12)
+  },
+  interactiveContainer: {
+    height: scale(40),
+    width: '100%',
+    alignItems: 'center'
+  },
+  dotContainer: {
+    height: '100%',
+    width: scale(190),
+    flexDirection: 'row',
+    justifyContent: 'space-between'
+  },
+  circle: {
+    borderWidth: 2,
+    borderColor: Colors.WHITE,
+    borderRadius: scale(15),
+    height: scale(30),
+    width: scale(30)
+  },
+  circleSected: {
+    backgroundColor: Colors.ACCENT_MINT,
+    borderWidth: scale(2),
+    borderColor: Colors.WHITE,
+    borderRadius: scale(15),
+    height: scale(30),
+    width: scale(30)
   }
 }
