+++ packages/edge-login-ui-rn/src/components/screens/existingAccout/ChangePinScreen.js
@@ -118,7 +118,7 @@ class ChangePinScreenComponent extends Component<Props, State> {
               </Text>
             </View>
             <View style={SetAccountPinScreenStyle.row2}>
-              <FourDigitInput style={SetAccountPinScreenStyle.fourPin} />
+              <FourDigitInput />
             </View>
             <View style={SetAccountPinScreenStyle.row3}>
               <Button
@@ -172,13 +172,6 @@ const SetAccountPinScreenStyle = {
     fontFamily: Constants.FONTS.fontFamilyRegular,
     color: Constants.GRAY_2,
     textAlign: 'center'
-  },
-  fourPin: {
-    ...Styles.FourDotInputDarkScaledStyle,
-    container: {
-      ...Styles.FourDotInputDarkScaledStyle.container,
-      height: scale(120)
-    }
   },
   nextButton: {
     upStyle: Styles.PrimaryButtonUpScaledStyle,
