+++ packages/edge-login-ui-rn/src/styles/index.js
@@ -10,7 +10,6 @@ export * from './common/InputStyles'
 export * from './common/ListItemStyles'
 export * from './common/LogoImageStyles'
 export * from './common/ModalStyles'
-export * from './components/FourDotInputStyle'
 export * from './components/OtpModalStyle'
 export * from './components/PasswordStatusStyle'
 export * from './ScreenStyle'
